+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Devop
 
+- Modularize store state [#1722](https://github.com/MyEtherWallet/MyEtherWallet/pull/1722)
 - Update packages from Greenkeeper [#1885](https://github.com/MyEtherWallet/MyEtherWallet/pull/1885)
 
 ### Release v5.2.11
@@ -108,6 +109,7 @@
 
 ### Devop
 
+- # Modularize store state [#1722][https://github.com/MyEtherWallet/MyEtherWallet/pull/1722]
 - Revert trezor version back to 7[#1737](https://github.com/MyEtherWallet/MyEtherWallet/pull/1772)
 - Refactor maker translations[#1737](https://github.com/MyEtherWallet/MyEtherWallet/pull/1737)
 - Add new MEW nodes[#1750](https://github.com/MyEtherWallet/MyEtherWallet/pull/1750)